Working singleton. Broken everything else.
