--- Assets/Code/Scripts/WeatherController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 997918f84c3ac894aabba983bf8caefb
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
