+++ Assets/InkLibrary.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 024a87478f1e90c4caf7e5ef59cec4e5
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
