+++ Assets/Level/Prefabs/EnvrionmentManager.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: bba5b2084342cab4087d1379d9de4a54
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
