--- Assets/Level/Prefabs/Rain Generator.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 3c53d464884d59146bfe655c04ada73e
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
