+++ Assets/Level/Prefabs/RainGenerator.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 65e60359046a57245baf1e483ad36d59
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
